+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## master
+## 1.2.0 (06.09.2018)
  + Add `httpClient`-Method to `LKDev\HetznerCloud\HetznerAPIClient`
  + Add `labels`-Property to `LKDev\HetznerCloud\Models\FloatingIp\FloatingIP`
  + Add `labels`-Property to `LKDev\HetznerCloud\Models\Server\Server`
@@ -16,7 +16,7 @@
  + You can now use the `labels`-Key on every `update`-Method, for easily updating the Labels
  + Add `LKDev\HetznerCloud\RequestOpts` - Class for easily customize the request opts. Could be used for filtering with the label selector.
  + Add the parameter `$requestOpts` to all `all`-Methods
- 
+--- 
 ## 1.1.0 (14.08.2018)
  + Add `update`-Method to `\LKDev\HetznerCloud\Models\Servers\Server`-Model for easily updateing the server meta data
    ```php